+++ changelog.md
@@ -411,6 +411,7 @@ Where $behavior is either rename or replace
 - Added sites resource for MsGraph
 - Added Pint with Laravel preset for formatting
 - Added composer scripts for formatting and testing
+- Added isaac/composer-git-hooks package to run scripts on commit
 
 ## Changed
 - Fixed an issue in MsGraph connect having an infinite loop when an error is returned from Microsoft